Improve modularity
