+++ app/client/src/pages/Editor/JSEditor/JSEditorToolbar/index.ts
@@ -1,3 +1,8 @@
 export { JSEditorToolbar } from "./JSEditorToolbar";
 
-export { type OnUpdateSettingsProps } from "./types";
+export {
+  type OnUpdateSettingsProps,
+  type JSActionDropdownOption,
+} from "./types";
+
+export { convertJSActionToDropdownOption, getJSActionOption } from "./utils";
