+++ app/client/src/pages/Editor/JSEditor/JSEditorToolbar/types.ts
@@ -1,7 +1,12 @@
 import type { JSAction } from "entities/JSCollection";
+import type { DropdownOption } from "@appsmith/ads-old";
 
 export interface OnUpdateSettingsProps {
   value: boolean | number;
   propertyName: string;
   action: JSAction;
+}
+
+export interface JSActionDropdownOption extends DropdownOption {
+  data: JSAction | null;
 }
